moved main to separate mock directory + now network is static library + moved flags sets to CMakeLists to enforce them when using project as a sub directory / dependency + now using std::filesystem::path in public functions + removed BASE_DIR macro + Logger improvements

+++ tests/MathTest.cpp
@@ -8,7 +8,7 @@
 
 #define EXPECT_EQ_FLOAT_VEC(expected, actual, thresh)           \
 do {                                                            \
-        for (size_t idx = 0; idx < expected.size(); ++idx)      \
+        for (uint32_t idx = 0; idx < expected.size(); ++idx)    \
             EXPECT_NEAR(expected[idx], actual[idx], thresh);    \
 } while (false)
